refactor(2025): Day 5 | p2
Doing everything in a singe pass instead.

Merging in a single pass instead of looping over and over again, and
after merging everything, summing up the finished range.

Doesn't actually save a lot of time as the execution time was already
trivial, but feels a lot nicer to look at.