refactor(2025): Day 2 | p2
Using simetry in repeating strings to detect them in O(n) instead of the
old O(n^2) solution.

Algorithm was found while searching for a more effective solution that
string manipilation.

According to chatgpt 4, using string manipulation is the quickest way in
js/ts to work on a digit level of longer numbers - so this is the
furtherst I'll optimize this.